moved audiodb api to new file, updated reviewroutes to display albums again, fixed add review button bug
